Reorganization of contract checkers structure
