+++ src/Aspect/Fetcher/ParentsContractsFetcher.php
@@ -8,7 +8,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpDeal\Aspect;
+namespace PhpDeal\Aspect\Fetcher;
 
 use Doctrine\Common\Annotations\Reader;
 use ReflectionClass;
